requirement.txt
ansible==2.9.5
asn1crypto==1.3.0
Babel==2.6.0
backports.csv==1.0.7
bcrypt==3.1.6
certifi==2019.9.11
cffi==1.12.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
defusedxml==0.6.0
diff-match-patch==20121119
Django==2.2.9
django-crispy-forms==1.8.0
django-formtools==2.1
django-import-export==1.2.0
django-ranged-response==0.2.0
django-redis==4.10.0
django-reversion==3.0.4
django-simple-captcha==0.5.11
djangorestframework==3.10.2
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.11.3
i18n==0.2
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
odfpy==1.4.0
paramiko==2.4.2
Pillow==6.2.0
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.4.2
PyNaCl==1.3.0
pytz==2019.1
PyYAML==5.1
redis==3.3.11
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tablib==0.12.1
tz==0.2.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.25.6
uWSGI==2.0.18
xlrd==1.2.0
xlwt==1.3.0
channels==2.3.1
channels-redis==2.4.1
celery==4.4.0
django-celery==3.3.1
django-celery-beat==1.6.0
django-celery-results==1.2.0

